requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langgraph>=0.0.20
pydantic>=2.5.0
tenacity>=8.2.0
tqdm>=4.66.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
prophet>=1.1.5
plotly>=5.18.0
matplotlib>=3.8.0
